Test Case Updated using Moq
