+++ GraduationTracker/GraduationTracker/GraduationTracker.csproj
@@ -40,14 +40,16 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Course.cs" />
-    <Compile Include="Diploma.cs" />
+    <Compile Include="IGraduationTracker.cs" />
+    <Compile Include="Model\Course.cs" />
+    <Compile Include="Model\Diploma.cs" />
     <Compile Include="GraduationTracker.cs" />
-    <Compile Include="STANDING.cs" />
+    <Compile Include="Model\STANDING.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Repository.cs" />
-    <Compile Include="Requirement.cs" />
-    <Compile Include="Student.cs" />
+    <Compile Include="Repository\IRepository.cs" />
+    <Compile Include="Repository\Repository.cs" />
+    <Compile Include="Model\Requirement.cs" />
+    <Compile Include="Model\Student.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
