+++ GraduationTracker/GraduationTracker/Model/Diploma.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GraduationTracker
+{
+    public class Diploma
+    {
+        public int Id { get; set; }
+        public int Credits { get; set; }
+        public int[] Requirements { get; set; }
+    }
+}
