+++ GraduationTracker/GraduationTracker/Model/Requirement.cs
@@ -1,0 +1,11 @@
+﻿namespace GraduationTracker
+{
+    public class Requirement
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public int MinimumMark { get; set; }
+        public int Credits { get; set; }
+        public int[] Courses { get; set; }
+    }
+}
