+++ GraduationTracker/GraduationTracker/Model/STANDING.cs
@@ -1,0 +1,12 @@
+﻿namespace GraduationTracker
+{
+    public enum STANDING
+    {
+        None,
+        SumaCumLaude,
+        MagnaCumLaude,
+        Average,
+        Remedial
+    }
+
+}
