--- GraduationTracker/GraduationTracker/STANDING.cs
@@ -1,12 +1,0 @@
-﻿namespace GraduationTracker
-{
-    public enum STANDING
-    {
-        None,
-        SumaCumLaude,
-        MagnaCumLaude,
-        Average,
-        Remedial
-    }
-
-}
